django/12일차(25.11.16)/practice/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = ""
authors = [
    {name = "choiguen",email = "daindj0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
